Restructure files and folders.
