+++ .gitignore
@@ -1,7 +1,7 @@
 # custom
 .idea
 data
-features/state_dicts
+src/features/state_dicts
 runs
 
 ### Python template
