+++ README.md
@@ -13,7 +13,7 @@ This will probably be turned into a Python package and moved to a new repository
 ### For CUB dataset
 1. Run `cub_download.py`.
 2. Run `cub_preprocess.py`.
-3. (Optional, but recommended) Download a [ResNet50 pretrained on iNaturalist2017](https://drive.google.com/drive/folders/1yHme1iFQy-Lz_11yZJPlNd9bO_YPKlEU) (filename on Google Drive: `BBN.iNaturalist2017.res50.180epoch.best_model.pth`) and place it in the folder `features/state_dicts`.
+3. (Optional, but recommended) Download a [ResNet50 pretrained on iNaturalist2017](https://drive.google.com/drive/folders/1yHme1iFQy-Lz_11yZJPlNd9bO_YPKlEU) (filename on Google Drive: `BBN.iNaturalist2017.res50.180epoch.best_model.pth`) and place it in the folder `src/features/state_dicts`.
 
 ### For development
 1. Install requirements from `requirements-dev.txt`.
