+++ requirements-dev.txt
@@ -1,1 +1,2 @@
-pytest>=7.2.2
+black~=23.3.0
+pytest~=7.2.2
