+++ src/prototree/project.py
@@ -5,8 +5,8 @@ import torch
 from torch.utils.data import DataLoader
 from tqdm import tqdm
 
-from prototree.models import ProtoTree
-from prototree.node import InternalNode, Node
+from src.prototree.models import ProtoTree
+from src.prototree.node import InternalNode, Node
 
 
 # TODO: generalize to ProtoBase
